requirements.txt
absl-py==1.1.0
aioredis==1.3.1
altair==4.2.0
anyio==3.3.1
anytree==2.8.0
apiclient==1.0.4
appdirs==1.4.4
argon2-cffi==21.1.0
arrow==1.2.2
asgiref==3.4.1
astroid==2.9.3
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
authentication==1.1.0
autobahn==22.2.2
Automat==20.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
binaryornot==0.4.4
bleach==4.1.0
blinker==1.4
boto3==1.17.93
botocore==1.20.112
branca==0.4.2
CacheControl==0.12.10
cachetools==5.0.0
camelcase==0.2
certifi==2020.11.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
clients==1.3
colorama==0.4.4
commonmark==0.9.1
comtypes==1.1.11
constantly==15.1.0
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.1
cv==1.0.0
cycler==0.11.0
daphne==3.0.2
debugpy==1.5.1
deco-helper==0.1.0
decorator==5.1.0
defusedxml==0.7.0rc1
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.13
django-3-jet==1.0.8
django-aadhaar==1.0.0.dev1
django-admin-csvexport==1.11
django-allauth==0.44.0
django-auth==0.1.9
django-browser-reload==1.6.0
django-ckeditor==6.2.0
django-cors-headers==3.3.0
django-cors-middleware==1.5.0
django-crispy-forms==1.10.0
django-csvexport==1.1.5
django-dropbox-storage==0.4.1
django-embed-video==1.4.2
django-extensions==2.2.5
django-filter==21.1
django-froala-editor==4.0.8
django-geoposition==0.3.0
django-heroku==0.3.1
django-jet==1.0.8
django-js-asset==2.0.0
django-location-field==2.1.0
django-model-utils==4.0.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-qr-code==3.0.0
django-social-auth==0.7.28
django-storages==1.11.1
django-suit==2.0a1
django-tailwind==3.3.0
django-templated-mail==1.1.1
django-toolbox==0.2
django-utils==0.0.2
django-utils-six==2.0
django-widget-tweaks==1.4.1
djangorestframework==3.11.0
djangorestframework-jsonapi==2.7.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
djangotoolbox==1.8.0
djoser==2.0.3
docutils==0.18.1
dropbox==11.28.0
easygui==0.98.2
ecdsa==0.17.0
entrypoints==0.3
et-xmlfile==1.1.0
face-recognition-models==0.3.0
fastapi==0.75.0
fcm-django==1.0.11
filelock==3.4.0
firebase-admin==5.2.0
Flask==2.0.2
flatbuffers==1.12
folium==0.12.1.post1
fonttools==4.33.3
future==0.18.2
gast==0.4.0
gcloud==0.18.3
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.27
gmane==0.1.dev25
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-currency==1.0.9
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.20
grpcio==1.45.0
grpcio-status==1.45.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.7.0
helper-functions==2.0.11
helpers==0.2.0
heroku==0.1.1
hiredis==2.0.0
httpcore==0.14.7
httpie==2.6.0
httplib2==0.20.4
httpx==0.22.0
hyperlink==21.0.0
idna==2.10
imageio==2.16.1
importlib-metadata==4.11.4
imutils==0.5.4
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonfield==3.1.0
jsonpath==0.82
jsonschema==3.2.0
jwcrypto==1.3.1
jwt==1.3.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
labelImg==1.8.6
lazy-object-proxy==1.7.1
libclang==14.0.1
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
msgpack==1.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
natsort==8.0.2
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.4
networkx==2.7.1
nltk==3.7
notebook==6.4.4
numpy==1.22.4
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.3
passlib==1.7.4
phonenumbers==8.12.25
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2021.11.23
platformdirs==2.4.0
plotly==5.6.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.24
proto-plus==1.20.3
protobuf==3.19.4
psycopg2==2.9.1
pusher==3.1.0
py==1.11.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.13.0
pydantic==1.9.0
pydeck==0.7.1
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.10.0
PyJWT==1.7.1
pylint==2.12.2
pymongo==4.0.2
Pympler==1.0.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
Pyrebase4==4.5.0
PyRect==0.2.0
pyrsistent==0.18.0
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.9
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.4
python-jose==3.3.0
python-jwt==3.3.2
python-multipart==0.0.5
python-slugify==6.1.2
python-social-auth==0.3.6
python3-openid==3.2.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pywin32==301
pywinpty==1.1.4
PyYAML==6.0
pyzbar==0.1.9
pyzmq==22.3.0
qrcode==7.3.1
qtconsole==5.1.1
QtPy==1.11.1
ratelim==0.1.6
razorpay==1.2.0
redis==4.1.2
regex==2022.6.2
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
rfc3986==1.5.0
rich==12.4.4
rsa==4.8
s3transfer==0.4.2
scikit-learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
secure-smtplib==0.1.1
segno==1.4.1
selenium==4.1.0
semver==2.13.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.15.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
social==0.1.dev3
social-auth-app-django==5.0.0
social-auth-core==4.0.3
sortedcontainers==2.4.0
soupsieve==2.3.1
South==1.0.2
SpeechRecognition==3.8.1
splinter==0.17.0
sqlparse==0.3.1
starlette==0.17.1
stone==3.3.1
streamlit==1.10.0
suit==2.0.2
tenacity==8.0.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.2
validate-email==1.3
validators==0.20.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
vonage==2.5.5
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
whitenoise==5.2.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.13.3
wsproto==1.0.0
WTForms==3.0.1
xgboost==1.6.1
xlrd==2.0.1
zipp==3.8.0
zope.interface==5.4.0
